requirements.txt
python-dotenv
faiss-cpu
langchain
requests
tqdm
bs4
streamlit